refactor: update Limine boot entry point and streamline request handling
