+++ kernel/src/boot/mod.rs
@@ -1,2 +1,1 @@
-#[cfg(target_arch = "riscv64")]
 pub mod limine;
